+++ CHANGELOG.md
@@ -1,8 +1,12 @@
 # CHANGELOG
 
+## master
+
+-   Implement #14: Add `stripTags` option.
+
 ## 2.1.0
 
--  Implement #12: Improve support for clipping HTML tables.
+-   Implement #12: Improve support for clipping HTML tables.
 
 ## 2.0.0